chore: moves testdata function to Go test file
